cipher/handlecipher: replace pubsub Subscription on Getter interface
In the future patches `DecryptCipherContext` will be used for
decrypting config having Publication instead of Subscription.
Inside decrypting functions the only method of the Subscription
interface is the `Get` method. In order to use both Publication
and Subscription the `Getter` interface was created.

In this patch the `Subscription` is replaced on `Getter` interface.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>