handleconfig: deprecate 'timer.use.config.checkpoint' usage
The default behaviour is to fetch saved config if its modification time
does not exceed some stale config time (timer.use.config.checkpoint).
This leads to the following problem observed by the customer:

   1. EVE node and an app are up running.
   2. Connection to the controller is lost for time > timer.use.config.checkpoint,
      so that last saved config modification time has not been updated.
   3. Node is rebooted, connection to the controller is still missing.
   4. Config is being fetched from a file, but modification time has not
      been updated for a while, thus config is rejected and applications
      do not show up.

Although the default setting for the timer is 7 days, it is not enough
for the node which constantly has problems with the internet and which
up-time is minimum, so it is not uncommon to have a last saved config
with the modification time dated several weeks from now.

This is not safe and having outdated config (basically this is always
the case, because there is always a time gap between actual config on
controller and config on EVE) is better, than to have no config at all.

This patch deprecates the 'timer.use.config.checkpoint' and config
is being always read from the file regardless of the modification time.

Kudos to Siddharth and Daniel for debugging this.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>