pillaer/zedcloud,zedagent: move auth container decryption to the zedagent
In case of the LOC the AuthContainer can come encrypted. When config is
received it is saved to the file in order to be read when there is no
connection to the controller.

This patch makes zedagent.handleconfig.go responsible for config decryption,
not the zedcloud as it was previously. Why? Config has to be saved after
decryption but before auth container verification, so that saved config
has correct verification signature.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>