pillar/zedagent: support handling of the compound edge device config
Compound config is the structure which includes main edge device
config in the protected envelope (auth container) and auxiliary
list of application/device commands like purge, reboot, shutdown
which are processed once main config is decrypted or/and verified
and successfully applied.

The compound config has the only use-case, which is the LOC, which
needs control over the edge device and all running applications.

Auxiliary commands which come along with the main edge device config
are processed by the functions, implemented for the LPS needs.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>