Move `persistcache` library from eve-libs to eve
This is part of configuration services implementation on EVE.
This library is required to store inline patch envelopes, since
they are defined as part of EdgeDevConfig and we want to serve
them as files.
After discussion in eve-libs repository
https://github.com/lf-edge/eve-libs/pull/5
It was decided that `persistcache` library will be moved to EVE
repo. Previous commits are using eve-libs version, after -- version
from EVE

Signed-off-by: Pavel Abramov <uncle.decart@gmail.com>