Do not suspend DPC testing when clocks are not synchronized
If DPC is failing connectivity tests (due to some transient network
issue) and device is powered off for long enough that clocks loose time
(or there is no battery to keep them running), device will not be able
to establish connectivity after power on for a very long time.
This is because we demand that failed DPC is tried again only after
some time elapses (5 minutes) since the last failed connectivity test.
However, when device starts with clocks reset at time "zero"
(i.e. start of the epoch), it will take many decades until the current
time (as reported by clocks) is past the last failure timestamp.
The solution is to detect unsync clocks and allow to test and use DPC
immediately.

Signed-off-by: Milan Lenco <milan@zededa.com>