debug/collect-info: attempt to install tar despite it is part of the debug container
In the commit 18f8fcfd1246 ("debug : add GNU tar to the container") the
gnu tar was droped from packages which should be installed in a favour to
the gnu tar package which has become a part of the debug container.
This is correct with one exception: old versions of EVE still don't
have gnu tar, so the script fails on old versions.

In order to be backward compatible still attempt to install gnu tar
which should be a NOP on latest EVE.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>