pillar/zedagent: repeat an attestation in case of a send failure
92c4e9d68063 ("Stale signing cert should not block sending device info")
introduced a problem of missing EVE certs on the controller side if very
first attestation request fails: specified commit ignores a send error of
the attestation request and does not repeat attestation.

The bug can be reproduced by applying iptables rules rejecting 443 port on
early boot, so that first send requests fail and attestation never happens
again.

The fix is simple: on a completion callback repeat attestation in case of
failure in 10 seconds delay to avoid tight fail-repeat loops.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>