Failover to normal SendOnIntf if nettrace fails to create HTTP client
In a rare case that nettrace fails to create HTTP client and establish
network tracing for whatever reason, it is more reasonable to simply
failover to running send operation without network tracing rather than
failing the request completely. After all, nettrace is not essential
and only helps with debugging of network issues.

Signed-off-by: Milan Lenco <milan@zededa.com>