Use 'defer' to simplify handling of AppNetworkStatus pending flags
Currently, we have to remember to set PendingAdd/PendingModify inside
AppNetworkStatus to false just before the corresponding handler returns.
However, there are few error branches where we forgot to clear
the pending flag. This causes zedmanager to just wait and not report
the error published inside the AppNetworkStatus.
Instead of fixing this case-by-case, let's take advantage of 'defer'
to make sure that we will never forget to clear the Pending flag,
even if a new branch with return statement is added in the future.

Signed-off-by: Milan Lenco <milan@zededa.com>