agentlog: remove stdout redirect
We used to redirect stdout for each individual service to a file in our
persistant storage, but this way these messages are not written into the
logs and it's easy to forget to check the file. By not redirecting the
messages we make sure all prints are available in our logs.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>