github/actions: Add GitHub action for auto-requesting code owner reviews.
This new GitHub Action automatically requests reviews from code owners when a
pull request is opened or updated. It fetches the necessary number of commits
related to the PR and identifies the appropriate code owners based on the
changed files. This is particularly useful as most of the code owners do not
have write access to the repository, and thus are not automatically assigned by
GitHub.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>