Revert "hypervisor/kvm: reduce VMM overhead to 100MiB for containers"
This reverts commit 6f761791b46394d5fa31377878f93ddd44baeb3c.

We notice increased number of customer issues on the latest EVE
versions, when QEMU process was killed by the OOM killer. So revert
the overhead to 600M value.

In the EVE mater the estimation algorithm is more sophisticated,
please take a look here:

   d88091ba7959 ("New formula to estimate VMM memory overhead (x86)")

The rule of thumb is to leave the overhead tuning to a customer using
the corresponding field on the controller side.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>