Add --fork option to 'eve exec'
Some commands, such as mmcli from the wwan container, need to create
child processes and expect that child and parent processes will be in
the same process namespace. However, if command with such expectations
is run under nsenter, it is necessary to fork after entering the
namespace, so that the parent process of the executed command is inside
the target namespace as well and not just its children.

Signed-off-by: Milan Lenco <milan@zededa.com>