xen-tools/chroot2: move chroot() and chdir() to a child process
There is no need to do chroot for the parent process before actual
clone is called, all the environment preparation should be as close
as possible to the actual execvp().

The other motivation for this change is the next patch, where PID
file of the child process will be created on the host, but not
in the chrooted container environment. So postpone chroot and move
it to the child process.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>