dom0-tools/bin/eve: add `list-app-consoles` and `attach-app-console` commands
The `list-app-consoles` command returns all running QEMU (KVM) consoles in the
following format:

  # /persist/eve list-app-consoles
  PID     APP-UUID                                CONS-TYPE       CONS-ID
  ---     --------                                ---------       ---------
  3883    e4e2f56d-b833-4562-a86f-be654d6387ba    VM              e4e2f56d-b833-4562-a86f-be654d6387ba.1.1/cons
  4072    f6d348cc-9c31-4f8b-8c4f-a4aae4590b97    CONTAINER       f6d348cc-9c31-4f8b-8c4f-a4aae4590b97.1.2/cons
  4072    f6d348cc-9c31-4f8b-8c4f-a4aae4590b97    VM              f6d348cc-9c31-4f8b-8c4f-a4aae4590b97.1.2/prime-cons

The `attach-app-console` command uses a console ID from the `list-app-consoles` as a parameter:

  # eve attach-app-console e4e2f56d-b833-4562-a86f-be654d6387ba.1.1/cons
  [20:26:15.116] tio v1.37
  [20:26:15.116] Press ctrl-t q to quit
  [20:26:15.116] Connected
  <PRESS ENTER>

  Ubuntu 18.04.6 LTS user hvc0

  user login:

The `attach-app-console` command attaches to the virtual QEMU console and pumps
bytes between socket and PTY, so that `tio` terminal can attach to the corresponding
PTY. This gives a fully working terminal of the guest.

Since `socat` and `tio` utilities exist only in the debug service EVE container,
the only way to use `attach-app-console` is to do the `eve enter debug` prior
the console attach. This is done by the `eve` script.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>