pillar/kvm: create another virtual console for the container only
This console is called a 'prime-cons' (could not think a better name)
and will connect the hosting Vm of a container with EVE environment.

So for a container there will be two virtual consoles: 'cons' and
'prime-cons'. The 'cons' is attached to the input and output of the
entry process of a container (this functionality exists for ages),
but a new 'prime-cons' is attached to the input and output of the
getty process of the hosting Vm.

The 'prime-cons' gives a way to execute commands on Vm of a
container.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>