xen-tools/init-initrd: start `agetty` in loop for the container Vm
Start `agetty` in loop on the '/dev/hvc1' virtual console (the other
end of this console is a 'prime-cons' visible on the EVE).

Since by default init process does not have any SID set (is 0), agetty
fails to set control terminal, so job control does not work. In order
to overcome this problem 'agetty' is called by the 'setsid' command,
which creates a SID.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>