GitHub Actions: use correct tag in assets.yml
We pass github.ref which have format refs/tags/%tag%. We can use this
format to download GitHub repository and checkout to specific branch,
however this format is not applicable to docker pull and rest of the
steps. Therefore we set up TAG environment variable in %tag% format
which is used in all steps after checking out and is derived from
git tag.

Signed-off-by: Pavel Abramov <uncle.decart@gmail.com>