wwan: try to find more useful reason for failure than DeadlineExceeded
When attempt to establish cellular connectivity fails, we want to
publish error that helps users to determine the root cause. Context
deadline exceeded is rarely usefull.
With this commit, mmagent tries to find the real reason behind the
timeout, and only returns the original DeadlineExceeded if it cannot
do any better.

Signed-off-by: Milan Lenco <milan@zededa.com>