grub: Add NVIVIA's devkit production module configuration
There is a version of the NVIDIA's Jetson Xavier NX developer kit that is
commercialized with a production module, which has eMMC, NVMe and doesn't
have SD Card slot. This commit adds the capability to detect those modules
(by reading the Asset-Tag from SMBIOS tables) and load the correct
parameters for boot (device tree, console, etc). The parameters for the
standard developer kit (with modules for development) are not affected, so
EVE should be able to boot on both variants without problems.

The Asset-Tag format was taken from the EEPROM layout description:
https://docs.nvidia.com/jetson/archives/r34.1/DeveloperGuide/text/HR/JetsonEepromLayout.html

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>