github/actions: In codeowners workflow fetch PR head ref for forked repos.
This commit adjusts the GitHub Action workflow to properly handle pull requests
from forked repositories. It removes the previous approach of calculating
fetch-depth and directly fetching the PR merge ref. Instead, it now checks out
the base ref with full history and then explicitly fetches and checks out the
PR's head ref. This change ensures that the workflow has access to the correct
set of changes, even when the PR is from a fork, and mitigates the potential
issue with the 'bad object' error seen when trying to access the head commit of
a forked repo.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>