pillar/qmp: do several attempts to retrieve valid QEMU status
QEMU status is very crucial to EVE and any error or unexpected status
leads to QEMU process is being stopped.

There is an issue in the 3rd QMP library:

   https://github.com/digitalocean/go-qemu/pull/210

And in order to be on a safe side and avoid these kind of problems in
the future repeat status qeury several times.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>