github/actions: exclude PR author from code owners review request.
The workflow has been updated to exclude the pull request author from the list
of potential reviewers. This change addresses a GitHub API error where review
requests failed if the author was also a code owner and thus included in the
review request payload. The fix iterates over the list of reviewers and removes
the PR author before constructing the JSON payload for the review request API
call.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>