pillar: Add new Makefile targets for Docker dev builds and shell access.
The Makefile has been updated to include two new targets: `build-docker-dev`
and `enter-docker-dev`. These targets are designed to facilitate the evaluation
of the codebase's state during the build process.

The `build-docker-dev` target constructs a Docker image up to the 'build'
stage, enabling developers to assess the intermediate state of the code,
particularly useful when applying patches that might only be present in
temporary containers and are difficult to inspect otherwise.

Following the build, the `enter-docker-dev` target allows for immediate shell
access within this intermediate Docker container, streamlining the process
for developers to interact with the code and build environment directly.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>