.github/workflows/eden.yml: remove report_pr_publish job
The report_pr_publish fails to comment on PRs because of some sort of
permission error, lets remove it for now. We still fail the Eden PR
test immidiatly if images are not published on dockerhub, so no worry
about waisting Eden runners looping in a confused state for 5 hours and
then failing.

Signed-off-by: Shahriyar Jalayeri <shahriyar@posteo.de>