usbmanager: fix read usb address for network cards
reading the usb address of the network device connected via usb
failed and set the UsbAddress of the ioBundle to an empty string;
at the same time it set the PCI address of the ioBundle to the
PCI address of the usb controller.
This made the usbmanager think it is a PCI passthrough rule for
this device and therefore usbmanager prevented all other usb devices
connected to that controller from passthrough.

Signed-off-by: Christoph Ostarek <christoph@zededa.com>