usbmanager: allow combination of usb{addr,product}
This allows the user to tighten the use of usb passthrough
even more.
It also makes it possible to only allow a bunch of specific
usb devices, but pass them through to different VMs.

E.g.: You have 3 keyboards from vendor X and 3 VMs.
Now it is possible to pass each of those keyboards to each VM
and at the same time only allow a passthrough of these keyboards
and therefore disallow any other device like a usb stick.

Signed-off-by: Christoph Ostarek <christoph@zededa.com>