pkg/xen-tools/chroot2: propagate error to a caller of chroot2 tool
Wait for a child exit status and returns it to a caller of a chroor2
caller.

Valid expectations are that init (process which is executed by the
following execvp() call) never exits, if it does - return an exit
status. It is possible to get 0 (success) of a chroot2 execution
status iif a spawned child process returned 0. If a child process
was terminated by a signal or coredumped (same as terminated by
a SIGSEGV), the chroot2 returns -1 (255).

Also patch replaces waitpid() on wait() for simplicity, since
chroot2 starts only one child there is no reason to specify
what child do we wait, wait() call waits for any child and is
similar in this context.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>