usbmanager: parse parentassigngrp
which will be used in later commits

Parent Assignment Group is there to reference the parent
assignment group in order to make the device dependent on
a different device.
Currently the concrete reason to do this is to make a usb
device dependent on the PCI address the USB controller is
using to prevent passthrough of the USB controller in one
application while trying to passthrough a USB device on
this controller to another application.

Inform the user that the parentassigngrp of their ioBundles
has problems:
- parentassigngrp is the same as the assigngrp
- different ioBundles with same assigngrp have different parentassinggrp
- loops in ioBundles

this makes it easier for the user to find the error in their
model manifest.

Signed-off-by: Christoph Ostarek <christoph@zededa.com>