usbmanager: allow vms with usb/pci clash
make it possible to have one vm with pci passthrough of
a usb controller while having another one with usb
passthrough of a device connected to this controller.

Obviously both vms cannot run in parallel and have
their passthroughs both working, but usbmanager will
try to do it's best to satisfy the customer by trying
to passthrough the usb device to the vm whenever possible;
this means that if the vm with pci passthrough is not running,
usbmanager will do a usb passthrough of the usb devices.

Several important changes have been made:

1. rule.priority now moved into eval:
  This is necessary as to return the correct priority in
  case composed rules are used
  With the compositionORRule the priority is dependent
  which passthrough rule is actually used.
  F.e. a compositionORRule could be composed of a
  usbPortPassthroughRule and a usbDevicePassthroughRule.
  The usbPortPassthroughRule has a higher priority than the
  usbDevicePassthroughRule, so when the overall priority is
  calculated it is important to take the priority of the
  actual rule, because otherwise the priority could be
  determined wrong and the device gets passed through
  to a different edge application then expected.

2. add ioBundleTree datastructure
  This datastructure enables to easily get depending
  ioBundles
  IOBundles that have the same assigngrp are put into
  an ioBundlesElem and the edges between the ioBundlesElems
  represent the parentassigngrp relation.
  This makes it possible to get all ioBundles that are dependent
  on an assigngrp, which is needed to update passthrough rules
  that are dependent on several ioBundles.

Signed-off-by: Christoph Ostarek <christoph@zededa.com>