pillar: Add warnings for unexpected total size changes in various modules.
This commit introduces additional logging warnings in the baseosmgr,
downloader, volumemgr, and related modules. The added warnings are triggered
when there is an unexpected change in the total size of a blob, specifically
when the total size is altered from a non-zero value to a different non-zero
value.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>
(cherry picked from commit c659aa5d762774596caa9b6b8f0ea97f769e39f7)