pillar: Enhance error logging in containerd client verification.
This commit adds a log entry in the `CtrGetImage` function to record exceptions
encountered during the verification of the containerd client. Additionally, it
introduces warning logs in the `verifyCtr` function to report when the
Containerd context is in an error state. These enhancements aim to improve the
visibility and traceability of errors related to client verification and
context handling in containerd operations.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>