Fix VLANs and LAGs
Support for VLANs and LAGs has been added to EVE quite some time ago,
but it has not been retested since. As we are soon planning
to productize this feature, this commit contains a set of patches
for regressions found while testing EVE with VLAN and LAG
configurations.

Patches are mostly about skipping L2-only ports and not trying to use
them for connectivity tests and HTTP requests.

Signed-off-by: Milan Lenco <milan@zededa.com>