pillar: Encode media type in verifier filenames.
Implemented a feature to save the media type hashed in the
verifier filename. This enhancement is used in both config and
status, ensuring that the media type is persistently available and
survives reboots. Modified several functions in verifier.go to
include media type in the filename generation. Also updated
related structures in verifiertypes.go to include the media type.

Issue: #3704

Signed-off-by: Avi Deitcher <avi@deitcher.net>