pillar/containerd: Avoid double quoted values for environment variables
When processing variables for containers defined at the controller
level, pillar creates the environment file with the proper shell
key/value pair, i.e. varname="value". However, if the value is already
enclosed in double quotes, the final result should be something like
varname=""value"", which will be misinterpreted by the shell will break
the shim VM initrd when initializing the container. This commit fixes
this issue by removing leading and trailing spaces and double quotes, so
values will not be quoted twice.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>