Enhance dmesg readability and add disk usage info in collect-info.sh
Modified the dmesg command to use the -T flag, which converts timestamps to
a human-readable format, improving log readability.

A new command, 'df -h', has been added to report on the filesystem disk space
usage in a human-readable format. This addition is crucial for a comprehensive
understanding of the filesystem's space utilization.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>