Disable image and blob GC in kubevirt eve
In kubevirt eve OCI images are downloaded by
1) eve downloader which downloads eve images and customer app images.
2) kubernetes downloader which downloads kubernetes components specific container images which are independent of eve.

But in both the cases we use same containerd repository to store the images.
Volumemgr runs GC when the device reboots and tries to delete all the images which are not referenced, and in the process
deleting all kubernetes images.

Until we find a reasonable solution to identify who downloaded the images like setting a label on eve downloaded images (eve-downloaded=true)
For now disabling GC on kubevirt eve.

Signed-off-by: Pramodh Pallapothu <pramodh@zededa.com>