pillar: Fix premature snapshot deletion in ZedManager.
This commit addresses an issue in the ZedManager where snapshots marked for
deletion were prematurely removed from the `SnapshotsToBeDeleted` list before
actual deletion could occur.

The fix involves appending snapshots to be deleted to the existing list instead
of overwriting it. This ensures that all snapshots intended for deletion are
retained in the list until they are actually deleted, preventing premature
removal from the list and ensuring consistent snapshot management.

The removal of the snapshot from the list of the snapshots to be deleted is
added to the triggerSnapshotDeletion function, so we remove it at the moment we
initiate the real deletion (when VM is halted).

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>