Remove interlock with controlled for EdgeNodeCerts
In
https://github.com/lf-edge/eve/commit/f40d6d43cb17101267507f2ee9be1e716409e6d3
and modified in
https://github.com/lf-edge/eve/commit/9ef0f0571d79812b6ac3b032542d78a2bd230683
EVE got an interlock to not fetch configuration from the controller
until after it has sent the EdgeNodeCerts. This was done since some
controller would send incomplete information when it could not do ECDH
and hence not do object encryption. That controller has now been fixed
(and is producing errors of the form "empty device certificate in create
cipher block method" in this case). Thus we can remove the interlock.

However, we still keep the retry logic from the second commit.

Signed-off-by: eriknordmark <erik@zededa.com>