Fix 'Skip triggering attest state machine before entering main loop'
The refactoring in
https://github.com/lf-edge/eve/commit/5832ab1c294e863b57915cb2fbb99da43a44070a
accidentally activated the subscriptions to AttestQuote and
EncryptedKeyFromDevice before the attest engine was started. As a result
we get the above log message and the encrypted storage key is not
reliably escrowed in the controller.

Here we start the attest engine before the subscriptions.

Signed-off-by: eriknordmark <erik@zededa.com>
(cherry picked from commit 6d8d99763e1c4914ce74e83d170bafe371a2dd99)