kvm: Introduce CAN Bus passthrough
This commit adds the capability to "passthrough" host CAN interfaces to
guests. The passthrough term is quoted because QEMU's implementation
actually emulates a CAN device to the Guest, connecting it to the
specificed CAN interface in the host.

Any Logical CAN defined in the device hardware model can be
connected to a Guest:

{
    "ztype": "IO_TYPE_LCAN",
    "phylabel": "can0.1",
    "logicallabel": "can0.1",
    "assigngrp" : "can0.1",
    "phyaddrs" : {
        "ifname" : "can0"
    },
}

The field 'ifname' is mandatory and corresponds to the host CAN interface
to be accessed by the Guest. A name for the emulated device should be given
in the phylabel and/or logicallabel fields. Having the former, priority
over the latter. Note that this name corresponds only to the QEMU's
configuration device name, it has no influence in the CAN interface's name
on the Guest, which will be OS dependent.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>