yetus: Install dependencies inside Yetus container
go-libzfs module depends on libzfs to be built. However, the library is
not available inside the Yetus container, leading to the following
error:

cmd/zfsmanager/handlediskconfig.go:12:9: could not import
github.com/bicomsystems/go-libzfs

This commit introduces the script tools/run-yetus.sh that allows to
install and manage all dependencies inside the container before start to
execute all the tests. The libzfs is built from sources in order to
match the version expected by pillar.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>