Add support for VNC console for shim VMs
We enable access to shim VMs via VNC console through the following
steps:
- add another virtual console in the QEMU configuration that's gonna be
used by VNC
- call agetty on that virtual console during VMs boot to enable login

When connecting to the QEMU instance via VNC the user will first be
presented with a view of the currently running process (container's
entry point). To switch to the shell of the shim VM the user will have
to press 'Ctrl+Alt+2' and then 'Enter' to get the login prompt.
Due to inability of some VNC clients to change the resolution of the
display, the client might crash. However upon the client restart the
new virtual console will appear as expected. The user can then switch
back to the previous virtual console by pressing 'Ctrl+Alt+1'.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>