pkg/*/make-raw: calcualate resulting image size dynamically
! Patch covers two almost identical 'make-raw' files.

The 'make-raw' script expects the resulting image being created by the
caller. The main problem of this approach is that the caller should
create an image in advance without knowing exactly the size of
partitions, which will be created lately. This causes a corruption of
the image with the ext4 rootfs, which is bigger than usual squashfs
and does not fit ~500Mb.

This patch fixes the problem by calculating resulting image size
by iterating over each requested partitions, e.g. 'imga' or 'imgb'
partitions are created by requesting size of the actual rootfs binary
image.

Having this calculations in place a caller does not need to predict
the resulting image size, but simply pass the '-C' argument, which
means the image file will be created according to the size of the
partitions.

Keeping the upgrade logic in mind we can't unfortunately change the
squashfs (default) rootfs partition size, so introduce minimum rootfs
size, which is equal to the default size of 512Mb.

We always use the maximum:

    max(512MB, file_size $ROOTFS_IMG)

so that default squashfs size won't be change, but ext4 will grow
(and yes, unfortunately you can't upgrade from squashfs to ext4,
 because ext4 partition is a way bigger)

Patch makes creation of the ext4 rootfs possible and also has a
size reduction for the installer.raw image:

                  before      after
 installer.raw     592Mb      296Mb

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>